Length composition object

An S4 object that holds length composition data in a specified format.

Arguments

title

A title for the object, useful for displaying the contents of the object

description

A longer description of the object

L_units

Units of measure for object, e.g., cm

L_type

Options are TL, FL, or SL (Total length, fork length, and standard length, respectively)

L_source

Options are FD or FI. FD is lengths or frequencies in the catch. FI is for fishery independent lengths or frequencies, such as a survey or research collection.

dataType

Options are Length or Frequency (See details for additional information)

observationGroup

Provide a description for interpreting column header, e.g., Year

header

Logical. Whether dataframe contains a descriptive header. FALSE will cause the header to be hidden

dt

The length composition data frame.

Details

The length composition object is contains a data frame of length observations. These observations are organized according to 'dataType'. When 'dataType' is set to 'lengths', each row contains a length observation and columns represent discrete observation units, such as year. When 'dataType' is set to 'freq', each row is the number of length observations corresponding to a length bin. The first column must contain the mid-points of each length bin, with columns 2 and above representing discrete observation units