Estimation of SPR using LBSPR method (Hordyk et al. 2016)

Hordyk, A.R., Ono, K., Prince, J.D., Walters, C.J., 2016. A simple length-structured model based on life history ratios and incorporating size-dependent selectivity: application to spawning potential ratios for data-poor stocks. Can. J. Fish. Aquat. Sci. 73, 1787–1799. https://doi.org/10.1139/cjfas-2015-0422

Usage

lbsprWrapper(
  LifeHistoryObj,
  LengthCompObj,
  Lc = 0,
  binWidth = 1,
  cvLinf = 0.1,
  byGroup = FALSE,
  modtype = "GTG"
)

Arguments

LifeHistoryObj

A LifeHistory object from fishSimGTG.

LengthCompObj

A LengthComp object

Lc

A minimum size limit in the same units of measure (L_units) and type of measurement (L_type) as the corresponding LengthComp object. Only used for analysis of fishery-independent length composition data.

binWidth

Used to create length-frequencies when LengthComp dataType is raw lengths.

cvLinf

Variability in Linf, used in LBSPR fitting. Default value same as that used in LBSPR library

byGroup

A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.

Examples

library(fishSimGTG)
library(LBSPR)
lbsprWrapper(fishSimGTG::LifeHistoryExample, fishLengthAssess::LengthCompExampleFreq)
#> A blank LB_pars object created
#> Default values have been set for some parameters
#> An object of class "LB_obj"
#> Slot "SPR":
#> [1] 0.3075997
#> 
#> Slot "Yield":
#> [1] 25583110901
#> 
#> Slot "YPR":
#> [1] 2572974
#> 
#> Slot "SSB":
#> logical(0)
#> 
#> Slot "SSB0":
#> logical(0)
#> 
#> Slot "B0":
#> logical(0)
#> 
#> Slot "LMids":
#>  [1]  11  13  15  17  19  21  23  25  27  29  31  33  35  37  39  41  43  45  47
#> [20]  49  51  53  55  57  59  61  63  65  67  69  71  73  75  77  79  81  83  85
#> [39]  87  89  91  93  95  97  99 101 103 105 107 109 111 113 115 117 119 121 123
#> [58] 125 127 129
#> 
#> Slot "pLCatch":
#>               [,1]
#>  [1,] 2.388294e-04
#>  [2,] 3.213268e-04
#>  [3,] 4.321033e-04
#>  [4,] 5.807221e-04
#>  [5,] 7.798918e-04
#>  [6,] 1.046436e-03
#>  [7,] 1.402514e-03
#>  [8,] 1.877106e-03
#>  [9,] 2.507760e-03
#> [10,] 3.342494e-03
#> [11,] 4.441613e-03
#> [12,] 5.878932e-03
#> [13,] 7.741505e-03
#> [14,] 1.012639e-02
#> [15,] 1.313243e-02
#> [16,] 1.684462e-02
#> [17,] 2.130940e-02
#> [18,] 2.650183e-02
#> [19,] 3.229074e-02
#> [20,] 3.841492e-02
#> [21,] 4.448769e-02
#> [22,] 5.004219e-02
#> [23,] 5.461316e-02
#> [24,] 5.782999e-02
#> [25,] 5.948596e-02
#> [26,] 5.955964e-02
#> [27,] 5.818836e-02
#> [28,] 5.561327e-02
#> [29,] 5.212036e-02
#> [30,] 4.799390e-02
#> [31,] 4.348826e-02
#> [32,] 3.881580e-02
#> [33,] 3.414601e-02
#> [34,] 2.961068e-02
#> [35,] 2.531221e-02
#> [36,] 2.133035e-02
#> [37,] 1.771149e-02
#> [38,] 1.447924e-02
#> [39,] 1.164202e-02
#> [40,] 9.195587e-03
#> [41,] 7.125154e-03
#> [42,] 5.407287e-03
#> [43,] 4.011759e-03
#> [44,] 2.903498e-03
#> [45,] 2.044595e-03
#> [46,] 1.396299e-03
#> [47,] 9.208688e-04
#> [48,] 5.831354e-04
#> [49,] 3.516739e-04
#> [50,] 1.995212e-04
#> [51,] 1.044402e-04
#> [52,] 4.878955e-05
#> [53,] 1.909450e-05
#> [54,] 5.429716e-06
#> [55,] 7.173350e-07
#> [56,] 0.000000e+00
#> [57,] 0.000000e+00
#> [58,] 0.000000e+00
#> [59,] 0.000000e+00
#> [60,] 0.000000e+00
#> 
#> Slot "pLPop":
#> numeric(0)
#> 
#> Slot "RelRec":
#> logical(0)
#> 
#> Slot "Ests":
#>       SL50  SL95   FM  SPR
#> [1,] 52.65 71.71 0.87 0.31
#> 
#> Slot "Vars":
#>          SL50     SL95         FM         SPR
#> [1,] 4.417234 10.52162 0.02189508 0.001982342
#> 
#> Slot "NLL":
#> [1] 98.17141
#> 
#> Slot "maxFM":
#> [1] 10
#> 
#> Slot "SPRatsize":
#> logical(0)
#> 
#> Slot "fitLog":
#> [1] 0
#> 
#> Slot "Species":
#> character(0)
#> 
#> Slot "MK":
#> [1] 1.5
#> 
#> Slot "M":
#> numeric(0)
#> 
#> Slot "Linf":
#> [1] 100
#> 
#> Slot "L_units":
#> character(0)
#> 
#> Slot "CVLinf":
#> [1] 0.1
#> 
#> Slot "L50":
#> [1] 66
#> 
#> Slot "L95":
#> [1] 67
#> 
#> Slot "Walpha":
#> [1] 0.01
#> 
#> Slot "Walpha_units":
#> [1] "g"
#> 
#> Slot "Wbeta":
#> [1] 3
#> 
#> Slot "FecB":
#> [1] 3
#> 
#> Slot "Steepness":
#> [1] 0.99
#> 
#> Slot "Mpow":
#> [1] 0
#> 
#> Slot "R0":
#> [1] 10000
#> 
#> Slot "SL50":
#> [1] 52.65
#> 
#> Slot "SL95":
#> [1] 71.71
#> 
#> Slot "MLL":
#> numeric(0)
#> 
#> Slot "sdLegal":
#> numeric(0)
#> 
#> Slot "fDisc":
#> numeric(0)
#> 
#> Slot "FM":
#> [1] 0.87
#> 
#> Slot "BinMin":
#> numeric(0)
#> 
#> Slot "BinMax":
#> numeric(0)
#> 
#> Slot "BinWidth":
#> numeric(0)
#> 
#> Slot "LData":
#>       [,1]
#>  [1,]    0
#>  [2,]    0
#>  [3,]    0
#>  [4,]    0
#>  [5,]    0
#>  [6,]    0
#>  [7,]    0
#>  [8,]    0
#>  [9,]    0
#> [10,]    1
#> [11,]    2
#> [12,]    5
#> [13,]    4
#> [14,]   12
#> [15,]   33
#> [16,]   65
#> [17,]   49
#> [18,]   52
#> [19,]   49
#> [20,]   55
#> [21,]   48
#> [22,]   63
#> [23,]   69
#> [24,]   60
#> [25,]   69
#> [26,]   74
#> [27,]   68
#> [28,]   82
#> [29,]   68
#> [30,]   85
#> [31,]   90
#> [32,]   59
#> [33,]   76
#> [34,]   55
#> [35,]   45
#> [36,]   27
#> [37,]   29
#> [38,]   23
#> [39,]    7
#> [40,]    8
#> [41,]    7
#> [42,]    6
#> [43,]    2
#> [44,]    2
#> [45,]    1
#> [46,]    0
#> [47,]    2
#> [48,]    2
#> [49,]    0
#> [50,]    0
#> [51,]    0
#> [52,]    0
#> [53,]    0
#> [54,]    0
#> [55,]    0
#> [56,]    0
#> [57,]    0
#> [58,]    0
#> [59,]    0
#> [60,]    0
#> 
#> Slot "Years":
#> [1] 1
#> 
#> Slot "NYears":
#> [1] 1
#> 
#> Slot "Elog":
#> [1] 0
#>