Utility that pools length data when multiple columns exist

Utility that pools length data when multiple columns exist

Usage

poolLengthComp(LengthCompObj, byGroup = FALSE)

Arguments

LengthCompObj

A life history object.

byGroup

A logical indicating whether quantity is to be calculated separately for each of multiple length comp groups (TRUE) or to length comp is to be pooled across groups prior to calculating quantity (default = FALSE). When TRUE, pooling is ignored if only a single group exists.

Examples

poolLengthComp(fishLengthAssess::LengthCompExampleFreq)
#>    Lmids Freq
#> 1     11    0
#> 2     13    0
#> 3     15    0
#> 4     17    0
#> 5     19    0
#> 6     21    0
#> 7     23    0
#> 8     25    0
#> 9     27    0
#> 10    29    1
#> 11    31    2
#> 12    33    5
#> 13    35    4
#> 14    37   12
#> 15    39   33
#> 16    41   65
#> 17    43   49
#> 18    45   52
#> 19    47   49
#> 20    49   55
#> 21    51   48
#> 22    53   63
#> 23    55   69
#> 24    57   60
#> 25    59   69
#> 26    61   74
#> 27    63   68
#> 28    65   82
#> 29    67   68
#> 30    69   85
#> 31    71   90
#> 32    73   59
#> 33    75   76
#> 34    77   55
#> 35    79   45
#> 36    81   27
#> 37    83   29
#> 38    85   23
#> 39    87    7
#> 40    89    8
#> 41    91    7
#> 42    93    6
#> 43    95    2
#> 44    97    2
#> 45    99    1
#> 46   101    0
#> 47   103    2
#> 48   105    2
#> 49   107    0
#> 50   109    0
#> 51   111    0
#> 52   113    0
#> 53   115    0
#> 54   117    0
#> 55   119    0
#> 56   121    0
#> 57   123    0
#> 58   125    0
#> 59   127    0
#> 60   129    0